Adaptive RandomWalk
Summary:
Defines `do_adaptation()` for `RandomWalkProposer`, based on the methods outlined in Garthwaite, Fan, and Sisson: https://www.tandfonline.com/doi/full/10.1080/03610926.2014.936562

Ten times during the adaptation period, the value `self.step_size` is tuned based on the samples collected since the last adaptation. The tuning equation is drawn from the work above, and is different in the univariate and multivariate cases.

Reviewed By: nazanint

Differential Revision: D20047117

fbshipit-source-id: 23c760df0781cd3925c3b668766bed0d0a584ce1