Remove query decorators when compiling Bean Machine to graph
Summary:
When we convert a Bean Machine program to one which, when executed, produces a graph, we no longer need the query decorator.

We still need to figure out whether we need to add a node to the graph indicating that the value in question is being queried, but we will do that in a later diff once I understand the semantics better.

Differential Revision: D20752021

fbshipit-source-id: da2a04191aac900210383742887d71fcd02cc2de