Add test case for BMA model to BM->BMG compiler
Summary:
I've added a test case that shows that we can compile a Bayesian Meta Analysis model into a graph. Note that this does not yet compile down to the actual BMG graph nodes; we just produce a graph builder that has the desired structure. We'll update this when the nodes are added to BMG proper.

The graph here looks quite complex but that is just because of the large number of samples; we have one sample for the mean of each of eight experiments, and then samples for the various sources of bias that we add together.

{F232770295}

Differential Revision: D20755741

fbshipit-source-id: 3b9cf10f983db06456c4587f0e7753adab8a29a2