Compiler now adds query nodes to graph
Summary:
A function decorated with the query decorator means "this function computes an intermediate value of interest, so track its value during inference". The compiler now transforms models into code which adds a query node to the graph.

The test case shows a highly simplified version of the Bayesian Meta Analysis model; the mean of the normal distribution is computed by a helper function and is a value we might be interested in, so we mark it in the graph by making a query node point to its associated computation node.

{F233025011}

Differential Revision: D20824231

fbshipit-source-id: 1d0e40e0ae4bba041d23b4cb837eb1992ffdfb86