Stop attempting to remove orphans from graphs
Summary:
In my original prototype for the compiler, we aggressively made everything in the program into a graph node, and then had to deal with a graph full of orphaned, useless nodes. So I wrote a helper method that removed the orphans.

The better technique is to simply not add nodes to the graph unless we know we will need them; a value stays a value until it becomes involved somehow in an operation with a sample from a distribution.

This then means that we no longer put orphaned nodes in the graph, so we do not need to remove them. I've deleted that step.

Unfortunately, this has a side effect. The algorithm which removed orphans also caused the order in which the graph nodes were enumerated to change slightly in some cases; since we are no longer running the orphan removal step, nodes in the graph are now being enumerated in the order that they were added, which changes the output of the graph visualizer.

I've fixed up all the test cases to match.

Differential Revision: D20806253

fbshipit-source-id: 840b82776ea44051d1287f41477c6699462a0dd1