Add `verbose` flag to `mh.infer()`
Summary:
This allows tqdm to be turned off, by passing a flag through `infer()` to `_infer()`.

I chose this approach of passing through the hierarchy because it should fare well in the future growth of BeanMachine's UX, where the `bool` should probably be replaced with several levels/layers of `verbose`, which might be set/modified at various layers of the inheritance hierarchy.

Conversely, this design might be too clunky/bloated because it passes through several layers of wrappers. Please let me know! An alternative approach could be to use some global variable. These are viable suggestions for this approach:
https://github.com/tqdm/tqdm/issues/619, https://github.com/tqdm/tqdm/issues/612

Reviewed By: nazanint

Differential Revision: D20741860

fbshipit-source-id: 73fbb5fa0543cf8331f6417f175442c482db9246