Second part of rewrite for comparison operations
Summary: The rewrite for comparison operators into SSA was done in two parts, and this is the second. The first part focused on binarization, that is, converting n-way comparisons into binary comparisons and binary conjunctions. This second part lifts out the first argument if it is a non-identifier, and if it is, lefts out the second argument. This enables the first rewrite.

Reviewed By: ericlippert

Differential Revision: D21032966

fbshipit-source-id: 88876aff8436b3fc7fb715dcb27e911f1d2bea49