Make observation node hold on to constant value
Summary:
In my original graph builder, an observation node points at the sample node it is observing, and a constant node containing the value. But this is unnecessary; logically, the observation node just holds the observed value; there's no need to make a special constant node that is not otherwise in the graph. Doing so just leads to a situation where we end up adding an unnecessary node into the BMG graph when we create it!

I've eliminated this problem; the observation node now holds on to the value directly rather than realizing it as a constant node, and we now do not generate a superfluous constant node in the BMG graph when we build it, or create a program which builds it.

Differential Revision: D21050700

fbshipit-source-id: 92e449623b9df8668d0933b0112a6b8e1c438b23