Minor changes to graph builder tests
Summary:
I've renamed some of the graph builder tests to have less generic names.

Also, in an upcoming diff I'm going to change how I output Python / C++ / graph nodes for the translation to BMG, which is of course what this is all about.

The nodes that we care about in an execution of BMG are the descendents of query, observation and sample nodes, so when I emit code or graph nodes, I'm going to strip out the nodes which are not a descendent of one of those.  This means that for the test case I've changed here, I need to add a query node "at the top" so that we keep the nodes in the output.

Differential Revision: D21077164

fbshipit-source-id: bc59bae0da22d95e8a94a7dcf8daa71da42dfe91