Re-implement orphan removal when traversing graphs
Summary:
In the original prototype of the compiler I ended up putting a lot of orphaned nodes into the graph. In a later revision the number of orphaned nodes was greatly reduced, so I deleted the "orphan removal" code.

We need it again because I am about to add some graph mutations that cause us to implement the BMG type system rules, and that will orphan some nodes.

The original orphan remover has not been restored because it had two undesirable qualities. First, it was destructive; it modified the graph. Second, it relied upon an invariant that I am *also* about to violate when I implement graph mutation. That is, we used to maintain the invariant that child nodes were always enumerated before parent nodes when walking the list of nodes, since they were added from leaves to roots. When we mutate the graph, there will be leaf nodes added later than parents.

I've therefore implemented a depth-first graph traversal that maintains the invariant that children are enumerated before parents. However, it is non-destructive, and it does an actual graph traversal from the roots rather than relying on a soon-to-be-unreliable implementation detail.

What are the roots?  BMG cares about three things: samples, queries and observations. We therefore take those nodes as roots, and do depth-first traversal from there.

Unfortunately this causes the nodes to be traversed in a slightly different order; previously they were traversed in the order they were added, since we knew that was always a valid order. Now they're traversed in depth first order.  That changes the output of some of the tests slightly.

Differential Revision: D21081702

fbshipit-source-id: 536be90fcb21a3d64c23cf21e8e59ffa22c62b1b