Restart end-to-end testing of BM->BMG compiler
Summary:
We've demonstrated that we can extract graphs from models, but due to type system restrictions and missing node types in BMG, I have  disabled all the tests for code paths where we generate BMG nodes, or emit programs that generate those nodes.

I've started to fix those type system problems starting with: Bernoulli distribution nodes that are not "logits" whose probabilities are
constant floats or tensors with a single float are now properly emitted as a BMG graph. That is, a simple coin flip, and nothing else,
now works end to end.

We still have a lot of work to do here: error reporting, making more cases work, accounting for the fact that Bernoulli results in a bool in BMG, not a tensor of floats, and so on, still needs to be implemented.

Also there appears to be a crashing bug in to_string, so I've disabled that test case until it is fixed.

Differential Revision: D20975319

fbshipit-source-id: 1152e50a8c59eed5db61d71b0a4648f25d09e46a