Refactoring _transform_list to allow it to handle other forms
Summary: The function _transform_list originally handled lists and later was extended to handle Tuples. In principle, it can be generalized to handle a variety of constructs that have a variable number of arguments. This diff refactors it to generalize it in a way that can enable testing such applications.

Reviewed By: ericlippert

Differential Revision: D21145896

fbshipit-source-id: cd5d0772ac182caae485ee9b761f42d8c94a6cce