Minor formatting cleanup in test output
Summary:
The graph builder can generate a program written in Python or C that when executed, builds the graph.

The way we output the Python code causes long lines, which is inconvenient for our test cases because it causes linter failures which we then have to work around.

I've fixed up the Python output to break up those long lines, matching what we were already doing for C output.

Differential Revision: D21163549

fbshipit-source-id: e4e77d3c92502418760c6ba97aa692c956eee7ed