Get result type of Bernoulli samples correct when type checking graph.
Summary:
I am continuing my slow but steady attempt to transform the graph which accurately represents the *Python* model into one that meets the constraints of the BMG type system.

In this diff, I solve a problem I've been mulling over for a while. In the Python-based graph, we represent a sample of a Bernoulli as being of the type that would actually be sampled from that Bernoulli in a Python program, which is typically a tensor of floats.  But what we want is to know that in the type system of the graph, a sample from a Bernoulli is of type bool, not float or tensor of floats.

The insight here is that things that are semantically different can be represented by different nodes. Rather than trying to mutate the existing Bernoulli node which represents the world as it was in the Python model, we can just replace it with a SimpleBernoulli node. The SimpleBernoulli node meets the constraints of the BMG type system:  it's sample type is bool, and it depends on a node of type `Probability`.

This change establishes that invariant in the graph; the next step is to support more cases where we ensure that the dependency can be made to be of type `Probability`.

This change once again caused the numbering scheme of the output to change, which caused some changes in test output.  I will probably soon make a change to how we assign these numbers so that this stops happening, as it is burdensome to have to keep updating the tests like this.

Differential Revision: D21167636

fbshipit-source-id: 952c4c2b23f891f20f902f63929932f4223edd80