custom type converter from AtomicValue to Python object
Summary: This diff makes it very easy to process the returned samples from inference as native Python types. Earlier these used to be of type AtomicValue which had to be inspected by calling `.bool` or `.real` to get the underlying value. Now these are automatically convereted to boolean or real as appropriate.

Differential Revision: D21214940

fbshipit-source-id: 408a2a4d3ea18fca5a5431a3751bc47f895d064b