Continue tweaking numbering of graph builder output.
Summary:
As in the previous diff in this stack, I've tweaked the code that generates a C++ program that in turn generates the graph so that the generated local variables are numbered in the order of the first statement in which they appear, not in the order they were constructed by the graph builder.

This should make the output easier to read and understand, but more importantly it will make it stable across changes to the graph builder implementation.

Differential Revision: D21267418

fbshipit-source-id: 89e63ad6a9e7f94bc4f4636d40021f5436f5c788