Small change to DOT output
Summary:
We order the DOT output in lexicographic order, nodes then edges -- because DOT doesn't care what order the nodes and edges come in. But we name the nodes N0, N1, ... and so on, which means that we sort them into order N0, N1, N10, N2... and this was bugging me. :)

I've fixed it so that we name the nodes N00, N01, N02, and so on, so that lexicographic order and number order are now the same.

Tiny change, lots of test changes.

Differential Revision: D21237054

fbshipit-source-id: d5d112d7d0cb93d88329467d82405fd7b02e567d