Another minor test output change
Summary:
One of the outputs of the compiler is a program which creates the graph; this program has a lot of "numbered" variables, and they are numbered in the order that the nodes were created.

In an earlier version of the compiler we maintained the invariant that child nodes, which must come before parent nodes, had smaller numbers, so the effect was that every line was number in ascending order.

We've violated that invariant in recent days, so the numbers are now in any old order. This is not just aesthetically vexing, it also presents a problem when changing the algorithm slightly, because the numbering of lots of nodes can change after small changes to how we transform the graph.

I'm going to re-establish the invariant that each statement in the program is associated with a unique, increasing integer.  This will change a lot of test cases once, but after that they should be more stable when we make small changes.

I've started with the python output.

Differential Revision: D21238402

fbshipit-source-id: 4fecd8f8e84598a048148d217361736ca1e0500b