Fix use-before-initialization problems in BMG
Summary:
When a graph node is created the type field of its associated value is set, but sometimes the value is not set. Unfortunately, the value can be read before it is set, for example by calling `to_string` on the graph. That's undefined behaviour, and our UB detector notices this and helpfully crashes.

To mitigate this, we now set the value to an appropriate default value when initializing the type.

Reviewed By: nimar

Differential Revision: D21282352

fbshipit-source-id: 6601a7b73ee1bb02bb5f22c59e782634dbff4c54