Enable end-to-end test on simple graph
Summary:
Now that the crashing bug in BMG `to_string` is fixed, I can finally once again do an end-to-end test that shows compiling a graph into a live BMG object that we can call methods on.

The "BAD" in the output is a bug in how BMG displays probabilities. I'll fix that in the next diff.

Reviewed By: nimar

Differential Revision: D21290069

fbshipit-source-id: fc2a799028bbee3f0f9faf2cf3cc6840d8ed97a0