Improve implementation of Bernoulli type fixing
Summary:
We accumulate a graph that correctly represents the type system of the user's Python model, which we must then transform into a graph that matches the Bean Machine Graph type system. My first attempt to "fix up" Bernoulli nodes involved creating a new "simplified" node, but upon reflection that was a mistake. We can simply mutate the node in place, and mark it as "fixed".

The invariants that we need to maintain are:

* If the Bernoulli node is marked as "logits" then its input probability must be a real number.
* If not, it must be a probability between 0.0 and 1.0.
* Either way, the sample type is bool.

I've started implementing this logic. However, note that if the model creates a "logits" Bernoulli parameterized by a sample, we automatically transform that into a "straight probability" Bernoulli. The test case I've added here demonstrates this.

In the next few diffs in this stack I will add a test case that demonstrates that the "logits" logic is correct.

Reviewed By: nimar

Differential Revision: D21312985

fbshipit-source-id: ce0c786a7e967143fe4a3821c224a3ba94c65d4e