End-to-end BM->BMG for beta distribution
Summary:
We accumulate a graph representing the original Python model with real-valued tensors for distributions, but must transform it into a graph that meets the BMG type system requirements.

In this diff I've added support for beta distributions; we ensure that the input types are positive reals and the output type is probability.

Reviewed By: nimar

Differential Revision: D21315670

fbshipit-source-id: 3caaad06b807c748f4dc76f775c141ce6a0a19e9