End-to-end BM->BMG for half Cauchy distribution
Summary:
We accumulate a graph representing the original Python model with real-valued tensors for distributions, but must transform it into a graph that meets the BMG type system requirements.

In this diff I've added support for half Cauchy distributions; we ensure that the input is a positive real and the output type is also a positive real.

Reviewed By: nimar

Differential Revision: D21316230

fbshipit-source-id: fc2bde14c89c194562a70d514ee00ac923e51870