fix an ergodicity issue with booleans
Summary: For uncoupled booleans NMC was proposing identical values because the current proposer for NMC for booleans was non-ergodic. An ergodic transition kernel has a  requirement to give a small mass to the current state of the chain. This is fixed here by changing the boolean proposers to give a 1% odds of staying in the current state.

Differential Revision: D21333207

fbshipit-source-id: 244476dad156aecc260e282d94c579961c82dc65