Renamed variable lhs to source
Summary: Initially lhs was meant to refer to the left hand side of the rewrite. Reading this a few weeks ago it was not clear lhs of what. For this reason, "source" seems like a better choice, as it might make it easier to recognize the value is the source term for the rewrite. Only time will show :-)

Reviewed By: ericlippert

Differential Revision: D21355582

fbshipit-source-id: 569e22520b11830641fdc03c4e9c96b6f98425bd