Add end-to-end support for binomial distributions in BM->BMG compiler
Summary:
We did not previously have any implementation of binomial distributions in the Beanstalk compiler; I have added end-to-end support in this diff.

However, I have not yet added computation of the support of an arbitrary torch binomial distribution, as that computation is a little tricky in general and will require more work. This means that we do not generate graphs for models where a binomial distribution is used to "index" another, like this contrived illustration:

    sample
    def flip(i):
      return Bernoulli(0.5)
    sample
    def bin():
      return Binomial(3, 0.5)
    sample
    def weird():
      return Normal(flip(bin()), 1.0)

There are four coin flips available: `flip(0)`, through `flip(3)`; the model chooses one of them by sampling a binomial; the result of the flip is then used as the mean of the normal.

To generate a graph we need to know how many coin flips are available so that there is one sample node for each. In this case it is easy; we could simply observe that the binomial distribution has a constant value for its first parameter. But if that constant were a tensor, or if it were stochastic, then we have a harder problem to solve, which we will do later.

Differential Revision: D21336617

fbshipit-source-id: f727aea028c330345801cfa62ff4b90d96a12b0c