Rewrite for lifting single regular argument
Summary: This rewrite is part of the set of rewrites that will be used to implement rewriting of function calls. In this case, we handle the case where there is a single regular argument and it is starred. This is the last rewrite that we need to get function calls to the normal form that we want to have by the time we are ready to pass it to BMG.

Reviewed By: ericlippert

Differential Revision: D21359014

fbshipit-source-id: 5d902d903be46c8f87e8ab938e1a765c59c80dc8