Implement if-then-else node in graph builder
Summary:
Nim has recently added an if-then-else node to BMG; I will use this when transforming the version of the graph that has captured the Python semantics into a graph which meets the BMG type system requirements.

For instance, if we have a model like:

    sample
    def flip():
      return Bernoulli(0.5)

    sample
    def weird():
      return Normal(flip() * 2.0, 1.0)

then in the Python version this is a multiplication of a tensor by a scalar, but in BMG that needs to be the multiplication of a real by a real.

A graph construction that satisfies BMG's type system would be to treat this as if we had `mult(if(sample, 1.0, 0.0), 2.0)`, or, if we feel like getting fancy, `if(sample, 2.0, 0.0)` to eliminate the multiplication entirely.

So far I've just defined the node type and made a trivial test case; I'll start implementing these sorts of transformations in the next diff.

A portion of the graph above is generated by the test case:

{F236208545}

Differential Revision: D21394490

fbshipit-source-id: e9b58e59c0175a624c01918c415665689431f292