Add disabled test for graph type conversion
Summary:
The next step in getting end-to-end BeanMachine to BMG compilation going is ensuring type conversions on samples are correct; for example, in Python we can use a sample from a Bernoulli -- a tensor 0.0 or 1.0 -- as the mean of a normal. But in BMG, those are different types -- bool vs real.

This is not yet implemented; I've checked in a disabled test case. It will be enabled in a few more diffs.

Differential Revision: D21415565

fbshipit-source-id: 8cc9b7892be282709c8e47b0bb47356e7442b846