Implement conversion of bool to real in BMG graph transformation
Summary: We accumulate a graph that describes the model but the node data types are those of the original Python program; we need to transform it into a graph that obeys the rules of the BMG type system.  In this diff we show that a conversion of a sample from a Boolean-typed distribution can be converted to a real number when required by the BMG type system.

Differential Revision: D21419113

fbshipit-source-id: 61d8ef6222e4c0e9f5dc421c8960455d2768aeb6