Add comments to graph builder
Summary:
Recent conversations with Detian and Walid have made it clear to me that the code in the graph builder is important, complicated, not very well orgaized, and very, very under-commented.

Since the graph builder is in some ways the core component of this compiler, it's important that it be not just correct, but also maintainable and understandable.

I've therefore started a process of:

* more clearly documenting the *purpose* of each class and each large section of the module
* more clearly identifying where work remains and what the work is -- it is good to know how much tech debt you're really in
* identifying regions of code that need to be cleaned up, refactored, or moved in later diffs.

I've been at it all day and I only made it two thirds of the way through the file; I'll do some refactorings and cleanup next, and then finish off by describing the type transformation algorithm in detail. Walid has raised excellent questions about how we know that it terminates, reaches a fixpoint, and so on; these deserve answers in the code.

Reviewed By: michaeltingley

Differential Revision: D21438596

fbshipit-source-id: cd3767ef7672b691b3e7bda9092d8c3662ba36f3