Move BMG node definitions to own module
Summary:
The Beanstalk compiler builds up a graph representation of a model by accumulating graph nodes into a graph builder as it executes. The module for the graph builder is getting rather large, and most of the "smarts" are not in the node definitions, but in sheer bulk of code the node definitions are getting unweildy.

Since the node definitions themselves do not in any way depend on the builder that accumulates them, we can move them into their own module (or modules -- we could further refactor into modules for operator nodes, distribution nodes, and so on, should we so choose.)

This is a good opportunity to move this code out of the utils folder and into a compiler folder.

No code changes here; I've just moved the code and slightly improved its ordering.

Differential Revision: D21465488

fbshipit-source-id: bb270cc461656909b085e3771c43a6c56c263d04