Rewrite rule for turing regular call alrguments into starred call arguments
Summary: This rewrite is part of a set of rewrites for implementing the processing of function calls. This rule looks for regular (that is, unstarred) arguments and turns them into starred arguments by adding *[...] around them. As such, it a step towrads turning multiple regular arguments into a single starred arugment.

Reviewed By: ericlippert

Differential Revision: D21509247

fbshipit-source-id: e49d921367fdee06628d5951759479fb9be5e3e9