fix an ergodicity issue with booleans
Summary: For uncoupled booleans NMC was proposing identical values because the current proposer for NMC for booleans was non-ergodic. An ergodic transition kernel is required to give a small mass to the current value of the state. This is fixed here by changing the boolean proposers to give a 1% odds of staying in the current state.

Differential Revision: D21333207

fbshipit-source-id: bcde20eddcd8bedc4a91ef62500e5373d5a8003d