Clarified error message "Expected success"
Summary: The original message is quite useful but does not provide a clear hint to a new user of the library that the cause of the problem is specifically that a pattern match has failed. Here the error message is expanded to be "Expected success but rewrite rule patten match failed".

Reviewed By: ericlippert

Differential Revision: D21525038

fbshipit-source-id: 3698febeaa8d250ada1364d4016b00373cc13ab4