fix an ergodicity issue with booleans
Summary: For uncoupled booleans NMC was proposing identical values because the current proposer for NMC for booleans was non-ergodic. An ergodic transition kernel is required to give a small mass to the current value of the state. This is fixed here by changing the boolean proposers to give a 1% odds of staying in the current state.

Reviewed By: dengdetian0603

Differential Revision: D21333207

fbshipit-source-id: be34e64fd6f5d50ac9bb258e1f8a03cc5db17ccc