Fix github build
Summary:
There is some sort of version mismatch causing the version of the ast module on the github build to reject usages of ast type constructors that have a mix of named and positional arguments. This fixes a couple of those that have crept in.

We'll solve the versioning problem in a later diff.

Differential Revision: D21587048

fbshipit-source-id: b48ec08122be6d06eaf65edb7f4ffb7d03c98339