Fix github build
Summary:
There is some sort of version mismatch causing the version of the ast module on the github build to reject usages of ast type constructors that have a mix of named and positional arguments. This fixes a couple of those that have crept in.

We'll solve the versioning problem in a later diff.

Differential Revision: D21587048

fbshipit-source-id: 0a8e50a963cf57450bffbe4a0a6ab89cca2663c8