Implement conversion from Boolean to natural in BMG type system
Summary:
We can now construct a BMG graph where a Boolean node (say, a sample from a Bernoulli) is used in a context where a natural number is expected -- say, the argument to a binomial.

I also fixed a bug along the way where binomial distributions with samples as a parameter were not being correctly emitted into the graph; this test also serves as a regressor for that bug.

Differential Revision: D21601708

fbshipit-source-id: e05e5874c1acf2458d365fda3b8bb3a3f6721ae4