Implement conversion from Boolean to natural in BMG type system
Summary:
We can now construct a BMG graph where a Boolean node (say, a sample from a Bernoulli) is used in a context where a natural number is expected -- say, the argument to a binomial.

I also fixed a bug along the way where binomial distributions with samples as a parameter were not being correctly emitted into the graph; this test also serves as a regressor for that bug.

Differential Revision: D21601708

fbshipit-source-id: af76b5c65158686e379e85408814e540e7979899