Implement conversion from Boolean to natural in BMG type system (#170)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/170

We can now construct a BMG graph where a Boolean node (say, a sample from a Bernoulli) is used in a context where a natural number is expected -- say, the argument to a binomial.

I also fixed a bug along the way where binomial distributions with samples as a parameter were not being correctly emitted into the graph; this test also serves as a regressor for that bug.

Reviewed By: nimar

Differential Revision: D21601708

fbshipit-source-id: 540a213a224f861f74a1f140df80bd79ac31dddc