Implement conversion of Boolean values to positive reals in BMG type system (#171)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/171

We now can generate correctly-typed graphs when a Boolean quantity, such as a sample from a Bernoulli, is used in a context expecting a positive real, such as the variance of a normal.

Two things to note, first, this is mathematically questionable. When treating a Boolean as a real value it should be either 0.0 or 1.0, but one of those is not a "positive real"; zero is a non-negative real.

Second, in service of the first point, as you can see in the output below, BMG silently corrects the error so that the zero value is treated as 0.0000000001.

Reviewed By: nimar

Differential Revision: D21600991

fbshipit-source-id: 4c9d9f227d19c4e1ee08519e1a50d04c1f1e92c3