Rewrite for lifting single keyword argument (#175)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/175

This rewrite is part of the set of rewrites that will be used to complete the rewriting of function calls by handling keyword arguments. In this case, we handle the case where there is a single keyword argument, which means it is a single argument that is double starred. This is the last rewrite that we need to get function calls to the normal form that we want to have by the time we are ready to pass it to BMG. As is noted in the body of the code, an example of this rewrite is to transform z = f(**({x:5}) into d={x:5}; z = f(*d). Technically, the rule requires that all processing of regular arguments is done. This is marked by having exacted one starred argument consisting of a variable. Also as explained in the comment in the new code, this rule is a slight generalization of what is needed, but this generalization a small optimization that is expected to be safe.

Reviewed By: ericlippert

Differential Revision: D21645745

fbshipit-source-id: d2fa7f902c37d2401029fdf1f4109b533278a710