Start fixing malformed addition and multiplication nodes by identifying them (#176)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/176

The BMG type system requires that the left and right inputs to a multiplication or addition be of the same type (real, positive real, and so on), and that the operand type is the output type.

When constructing a graph by executing a Python program, we record the actual Python types of the quantities, and fix them up to match the BMG type system rules after the accumulation is complete. In order to do so, an obvious first step is to identify which nodes are currently failing to meet the requirements. This diff adds a "malformed" node type; binary operator nodes with unequal input types are automatically marked as malformed.

Reviewed By: nimar

Differential Revision: D21646041

fbshipit-source-id: 94e6d6519986b7dee1f4f53f7e5025f92f6a257f