Rename @sample->@random_variable and @query->@functional (#178)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/178

As a beanmachine user, I should be able to easily understand the semantics from the name of bean machine decorators.

This diff renames `sample` to `random_variable` and `query` to `functional`, and also exposes the new names in the `beanmachine.ppl` top-level package. The previous names are flagged for deprecation and intentionally omitted from export.

Reviewed By: michaeltingley

Differential Revision: D21654220

fbshipit-source-id: e7d0eb9fcf27beb31e4a57caf69206acd0ecedd3