Tutorial - Estimating the Bias of a Coin (#206)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/206

In this tutorial we consider the problem of estimating the probability of heads for a potentially biased coin given a sample consisting of the result of tossing the same coin multiple times.

For reading convenience, this tutorial is published on ANP as n272056. This particular version is V24 and is available at https://www.internalfb.com/intern/anp/view/?id=272056&revision_id=186761939307211

Reviewed By: feynmanliang

Differential Revision: D21793942

fbshipit-source-id: 82c6e3224c1f740fe86cd80c969ca382a1656ed5