First draft of Bayesian Logistic Regression tutorial for documentation
Summary:
I've created a tutorial notebook describing how to do a tensorized Bayesian Logistic Regression model.

Thanks to Naz for help understanding the model and getting the code right.

A runnable version of this notebook is at:

https://www.internalfb.com/intern/anp/view/?id=280068

This is a first draft; I intend to add some additional code to use diagnostics.  Any additional comments, suggestions, improvements would be welcome.

Reviewed By: wtaha

Differential Revision: D22007816

fbshipit-source-id: 2ac7a6f1f025c94d6ec6b9d48be2bbf0289278a1