Start implementing type lattice logic (#215)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/215

After much thought I have arrived at an algorithm which should be able to convert a graph typed with "python" tensor types into an equivalent graph typed with Bean Machine Graph "semantic" types, like "probability" and "positive real".

The algorithm is most concisely expressed by describing the BMG type system as a *lattice*; that is a special kind of DAG which allows us to answer questions like "what is the most specific type that is more general than these two types?"

In lattice theory that function -- find the most specific thing more general than two other things -- is called "supremum", and I will need an implementation of it in my upcoming work.  This diff implements that function and a few tests.

Reviewed By: wtaha

Differential Revision: D22105249

fbshipit-source-id: b6612a50981751495abdfbb562327113603c657a