Continue implementing type inference algorithm
Summary:
I've devised a type inference algorithm for converting the accumulated graph, in which every node tracks the Python type, to valid BMG types.

A key component of this algorithm is what I'm calling the "infimum" type; that is, the *greatest lower bound type* in the type lattice for a given node. The infimum type is the *smallest* type that this node can be converted to.

In this diff I implement a property getter for the infimum type on all graph nodes. This includes some node types which we do not yet support in BMG. For those I've made a guess about what the type rules will be, but I have not implemented a test case. When we have defined rules for the new node types I will create tests.

Reviewed By: wtaha

Differential Revision: D22129159

fbshipit-source-id: b404240728cd902d72e1c566699162a4e2a512b5