Add a to-positive-real node to graph builder
Summary:
The BMG type system does not permit "implicit" conversions. For example, suppose we have a node with output, say "probability" and a different node with an input that requires a positive real -- say, the standard deviation of a normal.  We cannot use the probability node as an input, because a probability is not considered to be a kind of positive real. Rather, we must insert an explicit "to positive real" node.

This implements that node; the infimum type of "to positive real" is always a positive real.

Differential Revision: D22140430

fbshipit-source-id: f953cf9ef59bbd37e9745369b510fc98addc3bba