Add to-positive-real node construction in graph builder
Summary:
As noted in the previous diff, we will need the concept of "convert node to positive real" as a node in the graph when we convert an accumulated graph into its final BMG-compliant form.

This diff adds a memoized helper function to the graph builder to construct this node.

Reviewed By: wtaha

Differential Revision: D22142396

fbshipit-source-id: 7b6bc3659175363934f39087cfdddfdad8ae5670