Add to-positive-real node construction in graph builder (#221)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/221

As noted in the previous diff, we will need the concept of "convert node to positive real" as a node in the graph when we convert an accumulated graph into its final BMG-compliant form.

This diff adds a memoized helper function to the graph builder to construct this node.

Reviewed By: wtaha

Differential Revision: D22142396

fbshipit-source-id: 7e01a56e13201c5f4f5a5c4a61fa5070168f8710