Add input type requirements to node definitions
Summary:
Each node in a BMG-typed graph has type requirements on its input nodes; in this diff I make every node self-describe what its current requirements are in the given graph.

Some requirements  are *fixed and exact*; for example, a Normal requires that the standard deviation input be typed as Positive Real, and not some smaller type, like Probability, even though that would work.

Some requirements are *determined by context and exact*. For example, a multiplication of a Probability and a Positive Real will result in a Positive Real, but *also* requires that both its inputs have *exactly* the type Positive Real. In this case we will need to insert a ToPositiveReal node to make the graph type check.

Some requirements are *fixed and inexact*. For example, the input of a ToPositiveReal node must be any type *smaller than or equal to* Positive Real. (Right now we only need upper bounds; in the future we might also need the "any type larger than or equal to" constraint. When we do, I'll add the code.)

In this diff I just have every node compute its requirements, and provide a test case that checks every combination currently supported by both the Beanstalk compiler and BMG.

In upcoming diffs I will add code that looks for unresolvable violations, and code that looks for resolvable violations and fixes them.

Reviewed By: wtaha

Differential Revision: D22149020

fbshipit-source-id: c7ba49ab47a612d38a6a90c24bbbf2b7eba31ace