Add allowed list for safe functions in graph accumulation
Summary:
When accumulating the graph, we look for places where a function call is passed a graph node, so that for instance if we have a call to `exp`, we can intercept the call and rather than making it, generate a graph node.

We produce an error if an unsupported function is called with a graph node, but there are functions which we know are safe; for example, if someone is making a list, dictionary or set containing a graph node.

We now allow these; however, this will then introduce several new problems which have not yet been solved that we will have to consider:

* What if that list or dictionary is then in turn used as a "star argument" to one of our supported functions?
* What if the list or dictionary is passed as an argument to an unsupported function?
* What about predicates on lists like "is this element a member of this list?"

Reviewed By: wtaha

Differential Revision: D22313537

fbshipit-source-id: 7f066b777d9d227cc79738f8db55722ca46bcb33