Small modification to upper bound requirement construction
Summary:
When analyzing the accumulated graph to convert it to one supported by BMG, we put "requirements" on each edge; a requirement can be either an exact bound -- that is, the thing on the other end of the edge must be exactly of the given type -- or it can be an "upper bound" -- that is, the thing must be the given type or a smaller type.

It will be convenient in an upcoming diff for the helper method that creates upper bound requirement objects to be idempotent. I've made a small modification to it so that it can be passed an upper bound, and is then an identity.

Differential Revision: D22324435

fbshipit-source-id: b2eaf746a5ed0c7041fdf47157f6b1e0504a4d9c